rollback(api): remove subscriptions for now
