+++ graphql-server/src/resolvers.js
@@ -1,5 +1,6 @@
 const fs = require('fs').promises;
 const path = require('node:path');
+const replays = require('../data/replays')
 
 /*
     Get all players from data/users.txt
@@ -13,17 +14,6 @@ const getPlayers = async () => {
         const username = line.trim();
         return {username: username};
     })
-}
-
-/*
-    Set a new player in data/users.txt
-*/
-const setPlayer = async (username) => {
-    try {
-        await fs.appendFile(path.join('data', 'users.txt'), `\n${username}`);
-    } catch(error){
-        throw error;
-    }
 }
 
 /*
@@ -45,17 +35,6 @@ const getPlayer = async (username) => {
     return player;
 }
 
-const replays = [
-    {
-        "id": "replay-01",
-        "url": "https://static.wikia.nocookie.net/1d45d1a8-3e88-47e1-9e07-78f21f9d8ea7/scale-to-width/755",
-    },
-    {
-        "id": "replay-02",
-        "url": "https://img.buzzfeed.com/buzzfeed-static/static/2021-01/20/23/asset/9e2b6c32d7df/anigif_sub-buzz-20009-1611185881-21.gif",
-    },
-]
-
 const resolvers = {
     Query: {
         replays: () => replays.map(async (replay) => {
@@ -69,11 +48,10 @@ const resolvers = {
 
     Mutation: {
         addPlayer: async (_, {username}) => {
-            try{
-                await setPlayer(username);
-                return true;
-            } catch(_) {
-                return false;
+            try {
+                await fs.appendFile(path.join('data', 'users.txt'), `\n${username}`);
+            } catch(error){
+                throw error;
             }
         }
     }
